requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#

aiokafka==0.7.0           # via -r requirements.in
aiopg==1.1.0              # via -r requirements.in
aioredis==1.3.1           # via -r requirements.in, arq
arq==0.19.1               # via -r requirements.in
async-timeout==3.0.1      # via aioredis, arq
certifi==2020.12.5        # via httpx
click==7.1.2              # via arq
h11==0.11.0               # via httpcore
h2==3.2.0                 # via httpx
hiredis==1.1.0            # via aioredis
hpack==3.0.0              # via h2
httpcore==0.12.2          # via httpx
httpx[http2]==0.16.1      # via -r requirements.in
hyperframe==5.2.0         # via h2
idna==2.10                # via rfc3986
kafka-python==2.0.2       # via aiokafka
psycopg2-binary==2.8.6    # via -r requirements.in, aiopg
pydantic==1.7.3           # via arq
python-dotenv==0.15.0     # via -r requirements.in
rfc3986[idna2008]==1.4.0  # via httpx
sniffio==1.2.0            # via httpcore, httpx
